add OPLEM with siddon and phase-space i/o

+++ src/dev_c.h
@@ -41,7 +41,7 @@ void dev_MSPS_gen(float* msv, int nmsv, int* msi, int nmsi, int* nk, int nnk, in
 				  float* X, int sx, float* Y, int sy, float* Z, int sz, int* step, int nstep,
 				  int npoint, int seed, int nz, int ny, int nx);
 
-void dev_MSPS_acc(float* im, int nz, int ny, int nx,
+void dev_MSPS_acc(int* im, int nz, int ny, int nx,
 				  float* x, int sx, float* y, int sy, float* z, int sz);
 
 void dev_MSPS_naive(float* act, int nact, int* indact, int inact,
